requirements.txt
kubernetes==29.0.0
scapy==2.5.0
pyshark==0.6
python-dotenv==1.0.1